[I25-212] feat: Account 컴포넌트 개선 및 코드 정리
- text.css import 추가로 text-caption 클래스 적용 가능
- Account 컴포넌트에 사용자 프로필 이미지 표시 기능 추가
- GooglePopupLoginContent 중복 파일 삭제
- 불필요한 코드 제거 및 리팩토링
- 로그인 상태에서 프로필 이미지와 간소화된 UI 제공
