--- src/app/components/auth/GooglePopupLogin.tsx
@@ -1,92 +1,0 @@
-'use client';
-
-import { createClient } from '@/utils/supabase/client';
-import { useEffect, useState } from 'react';
-import { useModal } from '@/app/components/modal';
-import InputOfModal from '@/app/components/modal/inputs/InputOfModal';
-import { useProfileExistence } from '@/utils/hook/useProfile';
-import { profileConfig } from '@/services/config/profileConfig';
-
-function GooglePopupLoginContent() {
-  const supabase = createClient();
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const [hasShownModal, setHasShownModal] = useState(false);
-  const { openModal, closeModal } = useModal();
-  
-  const { data: profileExists, isLoading: isProfileLoading } = useProfileExistence();
-
-  const handleProfileSubmit = () => closeModal();
-
-  useEffect(() => {
-    const { data: { subscription } } = supabase.auth.onAuthStateChange(
-      async (event, session) => {
-        if (event === 'SIGNED_IN' && session) {
-          setIsLoggedIn(true);
-        } else if (event === 'SIGNED_OUT') {
-          setIsLoggedIn(false);
-          setHasShownModal(false);
-        }
-      }
-    );
-    return () => subscription.unsubscribe();
-  }, []);
-
-  useEffect(() => {
-    if (isLoggedIn && !isProfileLoading && profileExists === false && !hasShownModal) {
-      setHasShownModal(true);
-      openModal(
-        <InputOfModal
-          title="프로필 설정"
-          config={profileConfig}
-          onSubmit={handleProfileSubmit}
-        />
-      );
-    }
-  }, [isLoggedIn, isProfileLoading, profileExists, hasShownModal]);
-
-  const handleGoogleLogin = async () => {
-    const { data, error } = await supabase.auth.signInWithOAuth({
-      provider: 'google',
-      options: {
-        redirectTo: `${window.location.origin}/`
-      }
-    });
-
-  };
-
-  const handleLogout = async () => {
-    await supabase.auth.signOut();
-  };
-
-  if (isLoggedIn) {
-    return (
-      <div className="fixed top-4 right-4 z-50">
-        <button
-          onClick={handleLogout}
-          className="px-4 py-2 bg-red-500 text-white rounded-lg hover:bg-red-600 transition-colors"
-        >
-          로그아웃
-        </button>
-      </div>
-    );
-  }
-
-  return (
-    <div className="fixed top-4 right-4 z-50">
-      <button
-        onClick={handleGoogleLogin}
-        className="px-4 py-2 bg-blue-500 text-white rounded-lg hover:bg-blue-600 transition-colors flex items-center gap-2"
-      >
-        <svg width="20" height="20" viewBox="0 0 24 24">
-          <path fill="currentColor" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
-          <path fill="currentColor" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
-          <path fill="currentColor" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
-          <path fill="currentColor" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
-        </svg>
-        Google로 로그인
-      </button>
-    </div>
-  );
-};
-
-export default GooglePopupLoginContent;
