+++ src/app/globals.css
@@ -2,6 +2,8 @@
 @tailwind components;
 @tailwind utilities;
 
+@import './components/system/text.css';
+
 :root {
   --background: #f5f5f7;
   --foreground: #171717;
